pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "panoshu", email = "5207725+panoshu@users.noreply.github.com" },
]
requires-python = ">=3.13"
dependencies = [
    "doc2txt>=1.0.8",
    "fastapi[standard]>=0.135.2",
    "httpx>=0.28.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.3",
    "llama-index-core>=0.14.18",
    "llama-index-llms-openai-like>=0.7.1",
    "llama-index-readers-file>=0.6.0",
    "llama-index-vector-stores-milvus>=1.1.0",
    "pydantic>=2.12.5",
    "pydantic-settings[yaml]>=2.13.1",
    "pymilvus[milvus-lite,model]==2.6.9",
    "python-docx>=1.2.0",
    "setuptools<81",
    "spacy-pkuseg>=1.0.1",
    "streamlit>=1.55.0",
]

[project.scripts]
demo = "demo:main"

[build-system]
requires = ["uv_build>=0.10.8,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = ["httpx>=0.28.1", "pytest>=9.0.2"]
